powerpc/hvcall: define VASI signal opcode and signal codes
The H_VASI_SIGNAL hcall can be used by a suspending partition to
request the cancellation of a suspension in progress. Linux could use
this to cancel the operation while still running on the source system
if it encounters an unrecoverable issue (e.g. OOM).

Define the H_VASI_SIGNAL opcode and the associated signal codes. Only
the H_VASI_SIGNAL_CANCEL code will be used by Linux but all
architected signal codes are included for completeness.

Signed-off-by: Nathan Lynch <nathanl@linux.ibm.com>