powerpc/pseries/vasi_suspend_session: support new callbacks
Call the resume() callback once upon resuming, and call the the
complete() callback once the platform signals that the operation is
complete.

Callers can assume that their complete() callback will be called if
their resume() callback is called.

Add the appropriate expectations to the test code to verify the
intended behaviors.